Hoist UpdateStatus into its own file.
