+++ pixeltable/catalog/__init__.py
@@ -3,16 +3,7 @@
 from .catalog import Catalog
 from .column import Column
 from .dir import Dir
-from .globals import (
-    IfExistsParam,
-    IfNotExistsParam,
-    MediaValidation,
-    QColumnId,
-    RowCountStats,
-    UpdateStatus,
-    is_valid_identifier,
-    is_valid_path,
-)
+from .globals import IfExistsParam, IfNotExistsParam, MediaValidation, QColumnId, is_valid_identifier, is_valid_path
 from .insertable_table import InsertableTable
 from .named_function import NamedFunction
 from .path import Path
@@ -21,4 +12,5 @@ from .table import Table
 from .table_version import TableVersion
 from .table_version_handle import ColumnHandle, TableVersionHandle
 from .table_version_path import TableVersionPath
+from .update_status import RowCountStats, UpdateStatus
 from .view import View
