+++ pixeltable/catalog/table_version.py
@@ -29,14 +29,8 @@ if TYPE_CHECKING:
 
 from ..func.globals import resolve_symbol
 from .column import Column
-from .globals import (
-    _POS_COLUMN_NAME,
-    _ROWID_COLUMN_NAME,
-    MediaValidation,
-    RowCountStats,
-    UpdateStatus,
-    is_valid_identifier,
-)
+from .globals import _POS_COLUMN_NAME, _ROWID_COLUMN_NAME, MediaValidation, is_valid_identifier
+from .update_status import RowCountStats, UpdateStatus
 
 if TYPE_CHECKING:
     from pixeltable import exec, store
